feat: unify createFilesDocumentReferences and createDocumentReferences
